refactor: extract helpers into modules
